requirements.txt
# No external dependencies required - tkinter comes with Python 
Pillow>=10.0.0 